feat: add loading shimmer text component and tool activity step
- Introduced `LoadingShimmerText` component for loading states in chat.
- Implemented `ToolActivityStep` component to handle tool activity display, including status icons and details.
- Created utility functions for managing chat attachments, including preview URL handling and file size formatting.
- Developed composer palette items for slash commands, connections, and skills, enhancing user interaction in the chat composer.
- Added functions for parsing tool authorization and generating action summaries for various tool types.
- Implemented hooks for managing composer attachments and palette interactions, improving the overall user experience.
