+++ src/routes/Chat/index.tsx
@@ -5,16 +5,11 @@ import type {
   ChatContextMention,
   ChatMessage,
   ChatMessagePart,
-  ToolStatus,
 } from "../../../electron/chat/common.ts"
 import type { ConnectionProvider } from "../../../electron/connections/common.ts"
 import type { ModelChoice } from "../../../electron/models/common.ts"
-import type { ManagedSkillGroup } from "../../../electron/skills/common.ts"
 import type { AssistantTimelineBlock } from "./assistant-timeline.ts"
 import type { ChatTurn } from "./chat-turns.ts"
-import type { DraftAttachment } from "./composer-state.ts"
-import type { ComposerTrigger } from "./composer-triggers.ts"
-import type { ComposerPaletteItem } from "./ComposerPalette.tsx"
 import type { PromptInputMessage } from "@/components/ai-elements/prompt-input"
 import type { QueuedChatMessage } from "@/components/app-shell/chat-queue"
 import type { TranslateFn } from "@/i18n/i18n"
@@ -26,57 +21,38 @@ import {
   AlertTriangle,
   CheckIcon,
   ChevronRight,
-  Circle,
   CopyIcon,
   File as FileIcon,
-  FileArchive,
-  FileCode,
-  FileImage,
-  FilePenLine,
-  FilePlus2,
-  FileSearch,
-  FileSpreadsheet,
-  FileText,
-  FileVideoCamera,
   Folder,
-  FolderOpen,
-  Globe,
   ListChecks,
-  Loader2,
   Package,
   Plug,
   Plus,
-  PlayCircle,
-  Search,
-  SlidersHorizontal,
-  Square,
-  SquareTerminal,
   ThumbsDown,
   ThumbsUp,
-  Wrench,
   X,
 } from "lucide-react"
 import * as React from "react"
-import { toast } from "sonner"
 import { collectVisibleGeneratedArtifactSources } from "./artifact-sources.ts"
 import { splitAssistantTimelineBlocks, textFromTimelineBlocks } from "./assistant-timeline.ts"
+import { attachmentWithPreview } from "./chat-attachment-utils.ts"
 import { groupChatTurns, summarizeTurnProcess } from "./chat-turns.ts"
+import { AttachmentList } from "./ChatAttachments.tsx"
 import { ChatErrorNotice } from "./ChatErrorNotice.tsx"
-import { initialPaletteMode, resolveComposerPaletteKeyAction } from "./composer-palette-logic.ts"
+import { buildConnectionPaletteItems, buildSkillPaletteItems, slashCommandItems } from "./composer-palette-items.ts"
 import { composerReducer, contextMentionKey, initialComposerState } from "./composer-state.ts"
-import { detectComposerTrigger } from "./composer-triggers.ts"
 import { ComposerPalette } from "./ComposerPalette.tsx"
 import { ComposerTrailingControls } from "./ComposerTrailingControls.tsx"
+import { LoadingShimmerText } from "./LoadingShimmerText.tsx"
 import { assistantResponseActionTextByMessageId, copyableMessageText, visibleUserText } from "./message-text.ts"
 import { AddCustomModelDialog } from "./ModelControls.tsx"
 import { renderBlocks } from "./render-blocks.ts"
-import {
-  compactPathDetail,
-  compactToolDetail,
-  formatToolActivityDuration,
-  shouldShowRunningNoOutput,
-} from "./tool-activity.ts"
-import { hasStoppedTool, isToolCancellation } from "./tool-state.ts"
+import { formatToolActivityDuration } from "./tool-activity.ts"
+import { normalizeServiceSlug, toolActionSummary, toolServiceSlug } from "./tool-display.ts"
+import { hasStoppedTool } from "./tool-state.ts"
+import { ToolActivityStep } from "./ToolActivityStep.tsx"
+import { stripDraftAttachment, useComposerAttachments } from "./useComposerAttachments.ts"
+import { useComposerPalette } from "./useComposerPalette.ts"
 import { useModelCatalog } from "./useModelCatalog.ts"
 import { useVoiceComposerInput } from "./useVoiceComposerInput.ts"
 import { Conversation, ConversationContent, ConversationScrollButton } from "@/components/ai-elements/conversation"
@@ -95,9 +71,7 @@ import {
   PromptInputToolbar,
   PromptInputTools,
 } from "@/components/ai-elements/prompt-input"
-import { Shimmer } from "@/components/ai-elements/shimmer"
 import { Task, TaskContent, TaskTrigger } from "@/components/ai-elements/task"
-import { useChatService } from "@/components/AppContext"
 import { useSkillInventoryResource } from "@/components/AppDataHooks"
 import { Button } from "@/components/ui/button"
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
@@ -105,7 +79,6 @@ import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigge
 import { useT } from "@/i18n/i18n"
 import { cn } from "@/lib/utils"
 import { GeneratedArtifacts } from "@/routes/Chat/GeneratedArtifacts"
-import { ProviderIcon } from "@/routes/Connections/ProviderIcon"
 
 interface ChatAreaProps {
   billingCacheScope: string
@@ -134,559 +107,8 @@ interface ChatAreaProps {
   onViewBilling?: () => void
 }
 
-function stripDraftAttachment(attachment: DraftAttachment): ChatAttachment {
-  const { previewUrl: _previewUrl, ...chatAttachment } = attachment
-  return chatAttachment
-}
-
-interface AttachmentInput {
-  name: string
-  mime: string
-  size: number
-  path: string
-  kind?: "file" | "directory"
-  file?: File
-}
-
 const CHAT_CONTENT_MAX_WIDTH_CLASS = "min-w-0 max-w-[50rem]"
 const ASSISTANT_TEXT_SMOOTH_WINDOW_MS = 45_000
-const ATTACHMENT_PREVIEW_CACHE_LIMIT = 80
-
-const attachmentPreviewUrlByPath = new Map<string, string>()
-
-function revokePreviewUrl(url: string | undefined): void {
-  if (url?.startsWith("blob:")) {
-    URL.revokeObjectURL(url)
-  }
-}
-
-function setAttachmentPreviewUrl(path: string, url: string): void {
-  const current = attachmentPreviewUrlByPath.get(path)
-  if (current && current !== url) {
-    revokePreviewUrl(current)
-  }
-  if (!current && attachmentPreviewUrlByPath.size >= ATTACHMENT_PREVIEW_CACHE_LIMIT) {
-    const oldestPath = attachmentPreviewUrlByPath.keys().next().value as string | undefined
-    if (oldestPath) {
-      revokePreviewUrl(attachmentPreviewUrlByPath.get(oldestPath))
-      attachmentPreviewUrlByPath.delete(oldestPath)
-    }
-  }
-  attachmentPreviewUrlByPath.set(path, url)
-}
-
-function parseAuthorization(output: string | undefined): AuthorizationInfo | null {
-  if (!output) {
-    return null
-  }
-  try {
-    const parsed = JSON.parse(output) as Record<string, unknown>
-    if (
-      parsed.status === "authorization_required" &&
-      typeof parsed.service === "string" &&
-      typeof parsed.authUrl === "string"
-    ) {
-      return {
-        service: parsed.service,
-        displayName: typeof parsed.displayName === "string" ? parsed.displayName : parsed.service,
-        authUrl: parsed.authUrl,
-        message: typeof parsed.message === "string" ? parsed.message : undefined,
-      }
-    }
-  } catch {
-    return null
-  }
-  return null
-}
-
-function str(value: unknown): string {
-  return typeof value === "string" ? value : ""
-}
-
-function fileSizeLabel(size: number): string {
-  if (!Number.isFinite(size) || size <= 0) {
-    return ""
-  }
-  if (size < 1024) {
-    return `${size} B`
-  }
-  if (size < 1024 * 1024) {
-    return `${(size / 1024).toFixed(1)} KB`
-  }
-  return `${(size / 1024 / 1024).toFixed(1)} MB`
-}
-
-function attachmentExtension(name: string): string {
-  const lastSegment = name.split(/[\\/]/).pop() ?? name
-  const index = lastSegment.lastIndexOf(".")
-  return index > -1 ? lastSegment.slice(index + 1).toLowerCase() : ""
-}
-
-function isDirectoryAttachment(attachment: ChatAttachment): boolean {
-  return attachment.kind === "directory" || attachment.mime.toLowerCase() === "inode/directory"
-}
-
-function attachmentTypeLabel(t: TranslateFn, attachment: ChatAttachment): string {
-  if (isDirectoryAttachment(attachment)) {
-    return t("chat.attachmentFolder")
-  }
-  const extension = attachmentExtension(attachment.name)
-  if (extension) {
-    return extension.toUpperCase()
-  }
-  const [type] = attachment.mime.split("/")
-  return type ? type.toUpperCase() : "FILE"
-}
-
-function attachmentSummary(t: TranslateFn, attachment: ChatAttachment): string {
-  if (isDirectoryAttachment(attachment)) {
-    return attachmentTypeLabel(t, attachment)
-  }
-  const size = fileSizeLabel(attachment.size)
-  return size ? `${attachmentTypeLabel(t, attachment)} ${size}` : attachmentTypeLabel(t, attachment)
-}
-
-function isImageAttachment(attachment: ChatAttachment): boolean {
-  if (isDirectoryAttachment(attachment)) {
-    return false
-  }
-  if (attachment.mime.toLowerCase().startsWith("image/")) {
-    return true
-  }
-  return ["avif", "bmp", "gif", "jpeg", "jpg", "png", "svg", "webp"].includes(attachmentExtension(attachment.name))
-}
-
-function revokeAttachmentPreviewUrls(attachments: DraftAttachment[]): void {
-  for (const attachment of attachments) {
-    const cached = attachmentPreviewUrlByPath.get(attachment.path)
-    if (cached && (!attachment.previewUrl || cached === attachment.previewUrl)) {
-      revokePreviewUrl(cached)
-      attachmentPreviewUrlByPath.delete(attachment.path)
-    } else {
-      revokePreviewUrl(attachment.previewUrl)
-    }
-  }
-}
-
-function attachmentWithPreview(attachment: ChatAttachment): DraftAttachment {
-  if (!isImageAttachment(attachment)) {
-    return attachment
-  }
-  return {
-    ...attachment,
-    previewUrl: attachmentPreviewUrlByPath.get(attachment.path),
-  }
-}
-
-function filesFromDataTransfer(dataTransfer: DataTransfer): File[] {
-  const files = Array.from(dataTransfer.files)
-  if (files.length > 0) {
-    return files
-  }
-  return Array.from(dataTransfer.items)
-    .filter((item) => item.kind === "file")
-    .map((item) => item.getAsFile())
-    .filter((file): file is File => Boolean(file))
-}
-
-function hasKeys(value: Record<string, unknown> | undefined): boolean {
-  return Boolean(value && Object.keys(value).length > 0)
-}
-
-function normalizeServiceSlug(value: string): string {
-  return value.trim().replace(/^oo-/, "").toLowerCase()
-}
-
-function parseServiceFromCommand(command: string): string {
-  const serviceArg = String.raw`(?:"([^"]+)"|'([^']+)'|([A-Za-z0-9_-]+))`
-  const connectorMatch = command.match(
-    new RegExp(String.raw`(?:^|\s)(?:oo\s+)?connector\s+(?:schema|run)\s+` + serviceArg),
-  )
-  if (connectorMatch) {
-    return connectorMatch[1] ?? connectorMatch[2] ?? connectorMatch[3] ?? ""
-  }
-  const providerFlagMatch = command.match(new RegExp(String.raw`(?:--provider|--service)\s+` + serviceArg))
-  return providerFlagMatch ? (providerFlagMatch[1] ?? providerFlagMatch[2] ?? providerFlagMatch[3] ?? "") : ""
-}
-
-function toolServiceSlug(part: ChatMessagePart): string {
-  const input = part.input ?? {}
-  const fromInput = str(input.service)
-  if (fromInput) {
-    return normalizeServiceSlug(fromInput)
-  }
-  const auth = part.tool === "call_action" && part.status === "completed" ? parseAuthorization(part.output) : null
-  if (auth?.service) {
-    return normalizeServiceSlug(auth.service)
-  }
-  const skillTitle = part.title?.match(/^Loaded skill:\s*([A-Za-z0-9_-]+)/i)
-  if (skillTitle?.[1]) {
-    return normalizeServiceSlug(skillTitle[1])
-  }
-  const command = str(input.command)
-  if (command) {
-    const fromCommand = parseServiceFromCommand(command)
-    if (fromCommand) {
-      return normalizeServiceSlug(fromCommand)
-    }
-  }
-  return ""
-}
-
-function bashActionSummary(t: TranslateFn, command: string): string {
-  const normalized = command.replace(/\s+/g, " ").trim()
-  if (/^(ls|stat|file|du)\b/.test(normalized)) {
-    return t("chat.toolBashCheckFile")
-  }
-  if (/(^|[;&|]\s*)(which|command -v)\b|--version\b/.test(normalized)) {
-    return t("chat.toolBashCheckTools")
-  }
-  if (/^(sips|magick|convert|qlmanage)\b/.test(normalized)) {
-    return t("chat.toolBashConvertImage")
-  }
-  if (/^python3?\s+-c\s+["']import\b/.test(normalized)) {
-    return t("chat.toolBashCheckPythonModule")
-  }
-  if (/\bpip3?\s+install\b/.test(normalized)) {
-    return t("chat.toolBashInstallPythonPackage")
-  }
-  if (/^python3?\s+<<\s*['"]?EOF\b/.test(normalized) || /^python3?\s+\S+\.py\b/.test(normalized)) {
-    return t("chat.toolBashRunPythonScript")
-  }
-  if (/^(cat|sed|head|tail)\b/.test(normalized)) {
-    return t("chat.toolBashReadContent")
-  }
-  if (/^find\b/.test(normalized)) {
-    return t("chat.toolBashFindFiles")
-  }
-  return t("chat.toolRunGeneric")
-}
-
-/** 工具调用的一行人话动作摘要；原始命令只放在详情里。 */
-function toolActionSummary(t: TranslateFn, part: ChatMessagePart): string {
-  const input = part.input ?? {}
-  const service = str(input.service)
-  const action = str(input.action)
-  const target = service && action ? `${service} · ${action}` : service || action
-  const fallbackDetail = part.title || part.tool || "tool"
-  switch (part.tool) {
-    case "search_actions": {
-      const query = str(input.query)
-      return query ? t("chat.toolSearch", { detail: compactToolDetail(query) }) : t("chat.toolSearchGeneric")
-    }
-    case "inspect_action":
-      return target ? t("chat.toolInspect", { detail: target }) : t("chat.toolInspectGeneric")
-    case "call_action":
-      return target ? t("chat.toolCall", { detail: target }) : t("chat.toolCallGeneric")
-    case "bash": {
-      const command = str(input.command).split("\n")[0]
-      return command ? bashActionSummary(t, command) : t("chat.toolRunGeneric")
-    }
-    case "read": {
-      const filePath = str(input.filePath) || str(input.path)
-      return filePath ? t("chat.toolRead", { detail: compactPathDetail(filePath) }) : t("chat.toolReadGeneric")
-    }
-    case "write": {
-      const filePath = str(input.filePath) || str(input.path)
-      return filePath ? t("chat.toolWrite", { detail: compactPathDetail(filePath) }) : t("chat.toolWriteGeneric")
-    }
-    case "edit": {
-      const filePath = str(input.filePath) || str(input.path)
-      return filePath ? t("chat.toolEdit", { detail: compactPathDetail(filePath) }) : t("chat.toolEditGeneric")
-    }
-    case "list": {
-      const filePath = str(input.path) || str(input.filePath)
-      return filePath ? t("chat.toolList", { detail: compactPathDetail(filePath) }) : t("chat.toolListGeneric")
-    }
-    case "grep": {
-      const pattern = str(input.pattern)
-      return pattern ? t("chat.toolGrep", { detail: compactToolDetail(pattern) }) : t("chat.toolGrepGeneric")
-    }
-    case "glob": {
-      const pattern = str(input.pattern)
-      return pattern ? t("chat.toolGlob", { detail: compactToolDetail(pattern) }) : t("chat.toolGlobGeneric")
-    }
-    case "webfetch": {
-      const url = str(input.url)
-      return url ? t("chat.toolWebFetch", { detail: compactPathDetail(url) }) : t("chat.toolWebFetchGeneric")
-    }
-    case "task": {
-      return t("chat.toolTask", { detail: compactToolDetail(fallbackDetail) })
-    }
-    default:
-      return t("chat.toolGeneric", { detail: compactToolDetail(fallbackDetail) })
-  }
-}
-
-function toolStatusLabel(t: TranslateFn, status: ToolStatus | undefined): string {
-  switch (status) {
-    case "pending":
-      return t("chat.toolStatusPending")
-    case "running":
-      return t("chat.toolStatusRunning")
-    case "completed":
-      return t("chat.toolStatusCompleted")
-    case "error":
-      return t("chat.toolStatusError")
-    default:
-      return t("chat.toolStatusPending")
-  }
-}
-
-function toolPartStatusLabel(t: TranslateFn, part: ChatMessagePart): string {
-  return isToolCancellation(part) ? t("chat.toolStatusStopped") : toolStatusLabel(t, part.status)
-}
-
-function toolInlineDetail(part: ChatMessagePart): string {
-  if (part.tool !== "bash") {
-    return ""
-  }
-  const command = str(part.input?.command).split("\n")[0]
-  return command ? compactToolDetail(command, 96) : ""
-}
-
-function formatToolOutput(output: string | undefined): string {
-  if (!output) {
-    return ""
-  }
-  try {
-    return JSON.stringify(JSON.parse(output), null, 2)
-  } catch {
-    return output
-  }
-}
-
-function formatJson(value: Record<string, unknown>): string {
-  return JSON.stringify(value, null, 2)
-}
-
-function ToolStatusIcon({ status, stopped = false }: { status: ToolStatus | undefined; stopped?: boolean }) {
-  if (stopped) {
-    return <Square className="size-3.5 text-muted-foreground" />
-  }
-  switch (status) {
-    case "running":
-      return <Loader2 className="size-3.5 animate-spin text-muted-foreground" />
-    case "completed":
-      return <Circle className="size-3.5 text-muted-foreground" />
-    case "error":
-      return <AlertTriangle className="size-3.5 text-destructive" />
-    case "pending":
-    default:
-      return <Circle className="size-3.5 text-muted-foreground" />
-  }
-}
-
-function ToolActionIcon({ part }: { part: ChatMessagePart }) {
-  const className = "size-3.5 text-muted-foreground"
-  switch (part.tool) {
-    case "search_actions":
-      return <Search className={className} />
-    case "inspect_action":
-      return <SlidersHorizontal className={className} />
-    case "call_action":
-      return <PlayCircle className={className} />
-    case "bash":
-      return <SquareTerminal className={className} />
-    case "read":
-      return <FileText className={className} />
-    case "write":
-      return <FilePlus2 className={className} />
-    case "edit":
-      return <FilePenLine className={className} />
-    case "list":
-      return <FolderOpen className={className} />
-    case "grep":
-    case "glob":
-      return <FileSearch className={className} />
-    case "webfetch":
-      return <Globe className={className} />
-    case "task":
-      return <ListChecks className={className} />
-    default:
-      if (part.tool?.startsWith("todo")) {
-        return <ListChecks className={className} />
-      }
-      if (part.title?.match(/^Loaded skill:/i)) {
-        return <Package className={className} />
-      }
-      return <Wrench className={className} />
-  }
-}
-
-function ToolStepIcon({ part, provider }: { part: ChatMessagePart; provider?: ConnectionProvider }) {
-  const stopped = isToolCancellation(part)
-  if (provider && part.status !== "error" && !stopped) {
-    return <ProviderIcon iconUrl={provider.iconUrl} displayName={provider.displayName} size="compact" />
-  }
-  if (part.status === "error" || stopped) {
-    return <ToolStatusIcon status={part.status} stopped={stopped} />
-  }
-  return <ToolActionIcon part={part} />
-}
-
-function LoadingShimmerText({ children, className }: { children: string; className?: string }) {
-  return (
-    <Shimmer as="span" className={className} duration={2.4} spread={2.4}>
-      {children}
-    </Shimmer>
-  )
-}
-
-function ToolDetailSection({ label, children }: { label: string; children: React.ReactNode }) {
-  return (
-    <div className="space-y-1.5">
-      <div className="oo-text-micro font-medium text-muted-foreground uppercase">{label}</div>
-      {children}
-    </div>
-  )
-}
-
-function ToolPre({ children, tone = "default" }: { children: string; tone?: "default" | "error" }) {
-  return (
-    <pre
-      className={cn(
-        "oo-text-micro max-h-56 overflow-auto rounded-md border bg-background p-2.5 whitespace-pre-wrap",
-        tone === "error" && "border-destructive/25 bg-destructive/5 text-destructive",
-      )}
-    >
-      {children}
-    </pre>
-  )
-}
-
-function hasToolDetails(part: ChatMessagePart, auth: AuthorizationInfo | null): boolean {
-  const stopped = isToolCancellation(part)
-  return (
-    hasKeys(part.input) ||
-    hasKeys(part.metadata) ||
-    Boolean(part.output && !auth) ||
-    Boolean(part.error && !stopped) ||
-    Boolean(auth?.message) ||
-    shouldShowRunningNoOutput(part) ||
-    Boolean(part.attachmentsCount)
-  )
-}
-
-function ToolActivityStep({
-  part,
-  provider,
-  onAuthorize,
-}: {
-  part: ChatMessagePart
-  provider?: ConnectionProvider
-  onAuthorize: (auth: AuthorizationInfo) => void
-}) {
-  const t = useT()
-  const auth = part.tool === "call_action" && part.status === "completed" ? parseAuthorization(part.output) : null
-  const stopped = isToolCancellation(part)
-  const details = hasToolDetails(part, auth)
-  const statusText = toolPartStatusLabel(t, part)
-  const inlineDetail = toolInlineDetail(part)
-  const active = part.status === "pending" || part.status === "running"
-  const metaItems = [provider?.displayName, statusText].filter(Boolean)
-  const actionText = toolActionSummary(t, part)
-  const activeText = [actionText, inlineDetail, ...metaItems].filter(Boolean).join("  ")
-  const row = (
-    <div className="flex min-h-6 min-w-0 flex-1 items-center gap-2">
-      <span
-        className="flex size-5 shrink-0 items-center justify-center"
-        title={provider ? `${provider.displayName} · ${statusText}` : statusText}
-      >
-        <ToolStepIcon part={part} provider={provider} />
-      </span>
-      <div className="min-w-0 flex-1 overflow-hidden">
-        {active ? (
-          <div className="flex min-w-0 items-center">
-            <LoadingShimmerText className="min-w-0 truncate">{activeText}</LoadingShimmerText>
-          </div>
-        ) : (
-          <div className="flex min-w-0 items-center gap-2">
-            <span className={cn("min-w-0 truncate text-foreground", inlineDetail ? "shrink-0" : "flex-1")}>
-              {actionText}
-            </span>
-            {inlineDetail && (
-              <code className="min-w-0 flex-1 truncate rounded bg-muted px-1.5 py-0.5 font-mono text-[0.875em] text-muted-foreground">
-                {inlineDetail}
-              </code>
-            )}
-            <span className="flex min-w-0 shrink-0 items-center gap-1 text-muted-foreground">
-              {metaItems.map((item, index) => (
-                <React.Fragment key={`${index}:${item}`}>
-                  {index > 0 ? <span className="text-muted-foreground/70">·</span> : null}
-                  <span>{item}</span>
-                </React.Fragment>
-              ))}
-            </span>
-          </div>
-        )}
-        {auth && (
-          <div className="mt-1 flex flex-wrap items-center gap-2">
-            <span>{t("chat.authNeeded", { name: auth.displayName })}</span>
-            <Button size="sm" variant="outline" className="h-7 gap-1 px-2" onClick={() => onAuthorize(auth)}>
-              <Plug className="size-3.5" />
-              {t("chat.authorize")}
-            </Button>
-          </div>
-        )}
-      </div>
-    </div>
-  )
-
-  return (
-    <Collapsible defaultOpen={(part.status === "error" && !stopped) || Boolean(auth)}>
-      <div className="rounded-md">
-        {details ? (
-          <CollapsibleTrigger className="group/tool-step flex w-full items-center justify-between gap-2 text-left">
-            {row}
-            <ChevronRight className="size-3.5 shrink-0 text-muted-foreground opacity-0 transition-[opacity,transform] group-hover/tool-step:opacity-100 group-focus-visible/tool-step:opacity-100 group-data-[state=open]/tool-step:rotate-90 group-data-[state=open]/tool-step:opacity-100" />
-          </CollapsibleTrigger>
-        ) : (
-          row
-        )}
-      </div>
-      {details && (
-        <CollapsibleContent className="data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:animate-in data-[state=open]:fade-in-0">
-          <div className="ml-6 space-y-2.5 pt-1.5 pb-1">
-            {hasKeys(part.input) && (
-              <ToolDetailSection label={t("chat.toolParams")}>
-                <ToolPre>{formatJson(part.input ?? {})}</ToolPre>
-              </ToolDetailSection>
-            )}
-            {shouldShowRunningNoOutput(part) && (
-              <div className="oo-text-caption text-muted-foreground">{t("chat.toolRunningNoOutput")}</div>
-            )}
-            {part.output && !auth && (
-              <ToolDetailSection label={t("chat.toolResult")}>
-                <ToolPre>{formatToolOutput(part.output)}</ToolPre>
-              </ToolDetailSection>
-            )}
-            {part.error && !stopped && (
-              <ToolDetailSection label={t("chat.toolError")}>
-                <ToolPre tone="error">{part.error}</ToolPre>
-              </ToolDetailSection>
-            )}
-            {auth?.message && (
-              <ToolDetailSection label={t("chat.toolError")}>
-                <ToolPre tone="error">{auth.message}</ToolPre>
-              </ToolDetailSection>
-            )}
-            {hasKeys(part.metadata) && (
-              <ToolDetailSection label={t("chat.toolMetadata")}>
-                <ToolPre>{formatJson(part.metadata ?? {})}</ToolPre>
-              </ToolDetailSection>
-            )}
-            {part.attachmentsCount ? (
-              <div className="oo-text-caption text-muted-foreground">
-                {t("chat.toolAttachments", { count: part.attachmentsCount })}
-              </div>
-            ) : null}
-          </div>
-        </CollapsibleContent>
-      )}
-    </Collapsible>
-  )
-}
 
 type TurnProcessStatus = "running" | "completed" | "retrying" | "needsAction" | "error" | "stopped"
 
@@ -1378,230 +800,6 @@ const ChatTurnView = React.memo(function ChatTurnView({
   )
 }, chatTurnViewPropsEqual)
 
-function AttachmentPreviewTile({ attachment }: { attachment: DraftAttachment }) {
-  if (isDirectoryAttachment(attachment)) {
-    return (
-      <span className="flex size-10 shrink-0 items-center justify-center rounded-md bg-cyan-500/12 text-cyan-700 dark:text-cyan-300">
-        <Folder className="size-5" />
-      </span>
-    )
-  }
-
-  if (attachment.previewUrl && isImageAttachment(attachment)) {
-    return (
-      <span className="flex size-10 shrink-0 items-center justify-center overflow-hidden rounded-md bg-muted">
-        <img src={attachment.previewUrl} alt="" className="size-full object-cover" draggable={false} decoding="async" />
-      </span>
-    )
-  }
-
-  const mime = attachment.mime.toLowerCase()
-  const extension = attachmentExtension(attachment.name)
-
-  if (mime === "application/pdf" || extension === "pdf") {
-    return (
-      <span className="flex size-10 shrink-0 items-center justify-center rounded-md bg-red-500 text-[9px] font-semibold text-white">
-        PDF
-      </span>
-    )
-  }
-
-  const iconClassName = "size-5"
-  const tileClassName = "flex size-10 shrink-0 items-center justify-center rounded-md"
-
-  if (mime.startsWith("image/")) {
-    return (
-      <span className={cn(tileClassName, "bg-sky-500/12 text-sky-700 dark:text-sky-300")}>
-        <FileImage className={iconClassName} />
-      </span>
-    )
-  }
-  if (mime.startsWith("video/")) {
-    return (
-      <span className={cn(tileClassName, "bg-violet-500/12 text-violet-700 dark:text-violet-300")}>
-        <FileVideoCamera className={iconClassName} />
-      </span>
-    )
-  }
-  if (["zip", "gz", "tgz", "rar", "7z"].includes(extension)) {
-    return (
-      <span className={cn(tileClassName, "bg-amber-500/14 text-amber-700 dark:text-amber-300")}>
-        <FileArchive className={iconClassName} />
-      </span>
-    )
-  }
-  if (["csv", "tsv", "xls", "xlsx"].includes(extension)) {
-    return (
-      <span className={cn(tileClassName, "bg-emerald-500/12 text-emerald-700 dark:text-emerald-300")}>
-        <FileSpreadsheet className={iconClassName} />
-      </span>
-    )
-  }
-  if (["css", "html", "js", "json", "jsx", "md", "py", "ts", "tsx", "xml", "yaml", "yml"].includes(extension)) {
-    return (
-      <span className={cn(tileClassName, "bg-indigo-500/12 text-indigo-700 dark:text-indigo-300")}>
-        <FileCode className={iconClassName} />
-      </span>
-    )
-  }
-  if (mime.startsWith("text/") || ["doc", "docx", "rtf", "txt"].includes(extension)) {
-    return (
-      <span className={cn(tileClassName, "bg-muted text-muted-foreground")}>
-        <FileText className={iconClassName} />
-      </span>
-    )
-  }
-
-  return (
-    <span className={cn(tileClassName, "bg-muted text-muted-foreground")}>
-      <FileIcon className={iconClassName} />
-    </span>
-  )
-}
-
-function AttachmentImageCard({
-  attachment,
-  onOpen,
-  onRemove,
-}: {
-  attachment: DraftAttachment
-  onOpen: (attachment: DraftAttachment) => void
-  onRemove?: (id: string) => void
-}) {
-  const chatService = useChatService()
-  const [previewUrl, setPreviewUrl] = React.useState(attachment.previewUrl ?? null)
-  const attachmentPath = attachment.path
-  const attachmentMime = attachment.mime
-  const initialPreviewUrl = attachment.previewUrl ?? null
-  const imageAttachment = isImageAttachment(attachment)
-
-  React.useEffect(() => {
-    const cached = attachmentPreviewUrlByPath.get(attachmentPath) ?? initialPreviewUrl
-    setPreviewUrl(cached)
-    if (cached || !imageAttachment) {
-      return
-    }
-    let cancelled = false
-    void chatService
-      .invoke("getAttachmentPreview", { path: attachmentPath, mime: attachmentMime })
-      .then((result) => {
-        if (cancelled || !result.dataUrl) {
-          return
-        }
-        setAttachmentPreviewUrl(attachmentPath, result.dataUrl)
-        setPreviewUrl(result.dataUrl)
-      })
-      .catch(() => undefined)
-    return () => {
-      cancelled = true
-    }
-  }, [attachmentMime, attachmentPath, chatService, imageAttachment, initialPreviewUrl])
-
-  return (
-    <div className="group relative size-20 shrink-0">
-      <button
-        type="button"
-        title={attachment.path}
-        className="size-full overflow-hidden rounded-xl border border-border/60 bg-background text-left shadow-xs hover:border-border hover:bg-accent/40 focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 focus-visible:outline-none"
-        onClick={() => onOpen(attachment)}
-      >
-        {previewUrl ? (
-          <img
-            src={previewUrl}
-            alt=""
-            className="size-full object-cover object-center"
-            draggable={false}
-            decoding="async"
-          />
-        ) : (
-          <span className="flex size-full items-center justify-center text-muted-foreground/65">
-            <FileImage className="size-6" />
-          </span>
-        )}
-      </button>
-      {onRemove ? (
-        <button
-          type="button"
-          aria-label="Remove attachment"
-          className="absolute top-1 right-1 flex size-5 items-center justify-center rounded-full bg-foreground text-background shadow-sm hover:bg-foreground/85"
-          onClick={() => onRemove(attachment.id)}
-        >
-          <X className="size-3.5" />
-        </button>
-      ) : null}
-    </div>
-  )
-}
-
-function AttachmentList({
-  attachments,
-  className,
-  onRemove,
-}: {
-  attachments: DraftAttachment[]
-  className?: string
-  onRemove?: (id: string) => void
-}) {
-  const t = useT()
-  const chatService = useChatService()
-
-  const openAttachment = React.useCallback(
-    (attachment: DraftAttachment): void => {
-      void chatService.invoke("openLocalPath", { path: attachment.path }).catch((cause: unknown) => {
-        toast.error(t("chat.openAttachmentFailed", { error: cause instanceof Error ? cause.message : String(cause) }))
-      })
-    },
-    [chatService, t],
-  )
-
-  return (
-    <div className={cn("flex w-full flex-wrap justify-start gap-2", className)}>
-      {attachments.map((attachment) =>
-        isImageAttachment(attachment) ? (
-          <AttachmentImageCard
-            key={attachment.id}
-            attachment={attachment}
-            onOpen={openAttachment}
-            onRemove={onRemove}
-          />
-        ) : (
-          <div key={attachment.id} className="relative max-w-full min-w-0">
-            <button
-              type="button"
-              title={attachment.path}
-              className={cn(
-                "oo-border-divider flex h-14 max-w-full min-w-0 items-center gap-3 rounded-lg border bg-background/70 py-2 pl-2 text-left shadow-xs hover:border-border hover:bg-accent/60 focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 focus-visible:outline-none",
-                onRemove ? "pr-8" : "pr-2",
-              )}
-              onClick={() => openAttachment(attachment)}
-            >
-              <AttachmentPreviewTile attachment={attachment} />
-              <span className="min-w-0 flex-1">
-                <span className="block max-w-56 truncate text-sm leading-5 font-medium text-foreground">
-                  {attachment.name}
-                </span>
-                <span className="block truncate text-xs leading-4 font-normal text-muted-foreground">
-                  {attachmentSummary(t, attachment)}
-                </span>
-              </span>
-            </button>
-            {onRemove ? (
-              <button
-                type="button"
-                aria-label="Remove attachment"
-                className="absolute top-1/2 right-2 -translate-y-1/2 rounded-full p-1 text-muted-foreground hover:bg-muted hover:text-foreground"
-                onClick={() => onRemove(attachment.id)}
-              >
-                <X className="size-3.5" />
-              </button>
-            ) : null}
-          </div>
-        ),
-      )}
-    </div>
-  )
-}
-
 function contextMentionLabel(mention: ChatContextMention): string {
   return mention.kind === "skill" ? mention.name : mention.displayName
 }
@@ -1883,158 +1081,6 @@ const ChatTimeline = React.memo(function ChatTimeline({
   )
 })
 
-type SlashCommandAction = "billing" | "connections" | "insert" | "skills"
-
-interface SlashCommandPaletteItem extends ComposerPaletteItem {
-  action: SlashCommandAction
-  prompt?: string
-}
-
-interface ConnectionPaletteItem extends ComposerPaletteItem {
-  appId?: string
-  accountLabel?: string
-  displayName: string
-  service: string
-}
-
-interface SkillPaletteItem extends ComposerPaletteItem {
-  descriptionText: string
-  skillId: string
-  skillName: string
-}
-
-function normalizedSearchText(value: string): string {
-  return value.trim().toLowerCase()
-}
-
-function matchesComposerQuery(item: ComposerPaletteItem, query: string): boolean {
-  const normalized = normalizedSearchText(query)
-  if (!normalized) {
-    return true
-  }
-  return [item.id, item.title, item.description, item.meta ?? ""].some((value) =>
-    normalizedSearchText(value).includes(normalized),
-  )
-}
-
-function installedSkillHostCount(group: ManagedSkillGroup): number {
-  return group.hosts.filter((host) => host.status === "installed").length
-}
-
-function skillKindMeta(group: ManagedSkillGroup): string {
-  if (group.kind === "bundled") {
-    return "bundled"
-  }
-  if (group.kind === "registry") {
-    return "registry"
-  }
-  if (group.kind === "local") {
-    return "local"
-  }
-  return ""
-}
-
-function buildSkillPaletteItems(groups: ManagedSkillGroup[], fallbackDescription: string): SkillPaletteItem[] {
-  return groups
-    .filter((group) => installedSkillHostCount(group) > 0)
-    .slice()
-    .sort((left, right) => left.name.localeCompare(right.name))
-    .map((group) => ({
-      description: group.description || fallbackDescription,
-      descriptionText: group.description || fallbackDescription,
-      icon: <Package className="size-4" />,
-      id: `skill:${group.id}`,
-      meta: skillKindMeta(group),
-      skillId: group.id,
-      skillName: group.name || group.id,
-      title: group.name || group.id,
-    }))
-}
-
-function buildConnectionPaletteItems(
-  providers: ConnectionProvider[],
-  fallbackDescription: (service: string) => string,
-): ConnectionPaletteItem[] {
-  return providers
-    .filter((provider) => provider.status === "connected" && provider.appStatus === "active")
-    .slice()
-    .sort((left, right) => left.displayName.localeCompare(right.displayName))
-    .map((provider) => ({
-      accountLabel: provider.accountLabel,
-      appId: provider.appId,
-      description: provider.accountLabel || fallbackDescription(provider.service),
-      displayName: provider.displayName,
-      icon: <ProviderIcon iconUrl={provider.iconUrl} displayName={provider.displayName} size="compact" />,
-      id: `connection:${provider.service}:${provider.appId ?? "default"}`,
-      meta: provider.service,
-      service: provider.service,
-      title: provider.displayName,
-    }))
-}
-
-function slashCommandItems({
-  canViewBilling,
-  t,
-}: {
-  canViewBilling: boolean
-  t: TranslateFn
-}): SlashCommandPaletteItem[] {
-  return [
-    {
-      action: "skills",
-      description: t("chat.commandSkillsDescription"),
-      icon: <Package className="size-4" />,
-      id: "skills",
-      meta: "context",
-      title: t("chat.commandSkills"),
-    },
-    {
-      action: "connections",
-      description: t("chat.commandConnectionsDescription"),
-      icon: <Plug className="size-4" />,
-      id: "connections",
-      meta: "context",
-      title: t("chat.commandConnections"),
-    },
-    {
-      action: "billing",
-      description: t("chat.commandBillingDescription"),
-      disabled: !canViewBilling,
-      icon: <SlidersHorizontal className="size-4" />,
-      id: "billing",
-      meta: "ui",
-      title: t("chat.commandBilling"),
-    },
-    {
-      action: "insert",
-      description: t("chat.commandReviewDescription"),
-      icon: <FileSearch className="size-4" />,
-      id: "review",
-      meta: "prompt",
-      prompt: t("chat.commandReviewPrompt"),
-      title: t("chat.commandReview"),
-    },
-    {
-      action: "insert",
-      description: t("chat.commandSummarizeDescription"),
-      icon: <FileText className="size-4" />,
-      id: "summarize",
-      meta: "prompt",
-      prompt: t("chat.commandSummarizePrompt"),
-      title: t("chat.commandSummarize"),
-    },
-    {
-      action: "insert",
-      description: t("chat.commandStatusDescription"),
-      icon: <Circle className="size-4" />,
-      id: "status",
-      meta: "prompt",
-      prompt: t("chat.commandStatusPrompt"),
-      title: t("chat.commandStatus"),
-    },
-  ]
-}
-
 export function ChatArea({
   billingCacheScope,
   messages,
@@ -2061,9 +1107,7 @@ export function ChatArea({
   const modelCatalogState = useModelCatalog()
   const [composer, dispatchComposer] = React.useReducer(composerReducer, undefined, initialComposerState)
   const [inputError, setInputError] = React.useState<string | null>(null)
-  const fileInputRef = React.useRef<HTMLInputElement | null>(null)
   const textareaRef = React.useRef<HTMLTextAreaElement | null>(null)
-  const attachmentsRef = React.useRef<DraftAttachment[]>([])
   const appendVoiceTranscription = React.useCallback((text: string) => {
     dispatchComposer({ type: "append-transcription", text })
   }, [])
@@ -2076,12 +1120,12 @@ export function ChatArea({
   const composerDisabled = disabled || voiceInput.busy || initialSendPending
   const modelCatalog = modelCatalogState.catalog
   const modelError = modelCatalogState.error
-  const trigger = React.useMemo(
-    () => (composerDisabled ? null : detectComposerTrigger(draft, draftSelection.start, draftSelection.end)),
-    [composerDisabled, draft, draftSelection.end, draftSelection.start],
-  )
-  const triggerKey = trigger ? `${trigger.kind}:${trigger.start}:${trigger.query}` : null
-  const activeTrigger = triggerKey && triggerKey !== dismissedTriggerKey ? trigger : null
+  const composerAttachments = useComposerAttachments({
+    attachments,
+    disabled: composerDisabled,
+    dispatch: dispatchComposer,
+    setInputError,
+  })
   const slashItems = React.useMemo(
     () =>
       slashCommandItems({
@@ -2098,26 +1142,6 @@ export function ChatArea({
     () => buildConnectionPaletteItems(providers, (service) => t("chat.connectionFallbackDescription", { service })),
     [providers, t],
   )
-  const paletteItems = React.useMemo<ComposerPaletteItem[]>(() => {
-    if (!activeTrigger) {
-      return []
-    }
-    const sourceItems =
-      activeTrigger.kind === "skill" || paletteMode === "skills"
-        ? skillItems
-        : paletteMode === "connections"
-          ? connectionItems
-          : slashItems
-    return sourceItems.filter((item) => matchesComposerQuery(item, activeTrigger.query)).slice(0, 8)
-  }, [activeTrigger, connectionItems, paletteMode, skillItems, slashItems])
-  const paletteOpen = Boolean(activeTrigger)
-  const activePaletteItem = paletteItems[Math.min(activePaletteIndex, Math.max(0, paletteItems.length - 1))]
-  React.useEffect(() => {
-    attachmentsRef.current = attachments
-  }, [attachments])
-
-  React.useEffect(() => () => revokeAttachmentPreviewUrls(attachmentsRef.current), [])
-
   React.useEffect(() => {
     onArtifactsReset()
   }, [messages[0]?.id, onArtifactsReset])
@@ -2127,18 +1151,6 @@ export function ChatArea({
       onArtifactsReset()
     }
   }, [isGenerating, onArtifactsReset])
-
-  React.useEffect(() => {
-    dispatchComposer({ type: "set-active-palette-index", index: 0 })
-  }, [activeTrigger?.kind, activeTrigger?.query, paletteMode])
-
-  React.useEffect(() => {
-    if (!activeTrigger) {
-      dispatchComposer({ type: "set-palette-mode", mode: "root" })
-      return
-    }
-    dispatchComposer({ type: "set-palette-mode", mode: initialPaletteMode(activeTrigger.kind) })
-  }, [activeTrigger?.kind, activeTrigger?.start])
 
   const updateDraftSelection = React.useCallback(() => {
     const textarea = textareaRef.current
@@ -2174,137 +1186,21 @@ export function ChatArea({
     dispatchComposer({ type: "remove-context-mention", mention })
   }, [])
 
-  const returnToRootPalette = React.useCallback(() => {
-    const parentId = paletteMode === "connections" ? "connections" : "skills"
-    const parentIndex = slashItems.findIndex((item) => item.id === parentId)
-    dispatchComposer({ type: "set-palette-mode", mode: "root" })
-    dispatchComposer({ type: "set-active-palette-index", index: parentIndex >= 0 ? parentIndex : 0 })
-  }, [paletteMode, slashItems])
-
-  const applySlashCommand = React.useCallback(
-    (item: SlashCommandPaletteItem, currentTrigger: ComposerTrigger) => {
-      if (item.disabled) {
-        return
-      }
-      if (item.action === "skills") {
-        dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement: "/" })
-        dispatchComposer({ type: "set-palette-mode", mode: "skills" })
-        focusDraftAt(currentTrigger.start + 1)
-        return
-      }
-      if (item.action === "connections") {
-        dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement: "/" })
-        dispatchComposer({ type: "set-palette-mode", mode: "connections" })
-        focusDraftAt(currentTrigger.start + 1)
-        return
-      }
-      if (item.action === "billing") {
-        dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement: "" })
-        onViewBilling?.()
-        focusDraftAt(currentTrigger.start)
-        return
-      }
-
-      const replacement = `${item.prompt ?? ""} `
-      dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement })
-      focusDraftAt(currentTrigger.start + replacement.length)
-    },
-    [focusDraftAt, onViewBilling],
-  )
-
-  const applySkillItem = React.useCallback(
-    (item: SkillPaletteItem, currentTrigger: ComposerTrigger) => {
-      addContextMention({
-        description: item.descriptionText,
-        id: item.skillId,
-        kind: "skill",
-        name: item.skillName,
-      })
-      dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement: "" })
-      focusDraftAt(currentTrigger.start)
-    },
-    [addContextMention, focusDraftAt],
-  )
-
-  const applyConnectionItem = React.useCallback(
-    (item: ConnectionPaletteItem, currentTrigger: ComposerTrigger) => {
-      addContextMention({
-        ...(item.accountLabel ? { accountLabel: item.accountLabel } : {}),
-        ...(item.appId ? { appId: item.appId } : {}),
-        displayName: item.displayName,
-        kind: "connection",
-        service: item.service,
-      })
-      dispatchComposer({ type: "replace-trigger", trigger: currentTrigger, replacement: "" })
-      focusDraftAt(currentTrigger.start)
-    },
-    [addContextMention, focusDraftAt],
-  )
-
-  const applyPaletteItem = React.useCallback(
-    (item: ComposerPaletteItem | undefined) => {
-      if (!item || !activeTrigger) {
-        return
-      }
-      if (activeTrigger.kind === "slash" && paletteMode === "root") {
-        applySlashCommand(item as SlashCommandPaletteItem, activeTrigger)
-      } else if (paletteMode === "connections") {
-        applyConnectionItem(item as ConnectionPaletteItem, activeTrigger)
-      } else {
-        applySkillItem(item as SkillPaletteItem, activeTrigger)
-      }
-    },
-    [activeTrigger, applyConnectionItem, applySkillItem, applySlashCommand, paletteMode],
-  )
-
-  const handleComposerKeyDown = React.useCallback(
-    (event: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      if (event.nativeEvent.isComposing) {
-        return
-      }
-      if (!paletteOpen) {
-        return
-      }
-      const action = resolveComposerPaletteKeyAction({
-        activeIndex: activePaletteIndex,
-        activeRootAction:
-          activeTrigger?.kind === "slash" && paletteMode === "root" && activePaletteItem
-            ? (activePaletteItem as SlashCommandPaletteItem).action
-            : undefined,
-        itemCount: paletteItems.length,
-        key: event.key,
-        paletteMode,
-        triggerKind: activeTrigger?.kind,
-      })
-      if (action.type === "none") {
-        return
-      }
-      event.preventDefault()
-      if (action.type === "move") {
-        dispatchComposer({ type: "set-active-palette-index", index: action.index })
-      } else if (action.type === "back") {
-        returnToRootPalette()
-      } else if (action.type === "open-root-item" && activeTrigger && activePaletteItem) {
-        applySlashCommand(activePaletteItem as SlashCommandPaletteItem, activeTrigger)
-      } else if (action.type === "select") {
-        applyPaletteItem(activePaletteItem)
-      } else if (action.type === "dismiss") {
-        dispatchComposer({ type: "set-dismissed-trigger-key", key: triggerKey })
-      }
-    },
-    [
-      activePaletteItem,
-      activePaletteIndex,
-      activeTrigger,
-      applyPaletteItem,
-      applySlashCommand,
-      paletteItems.length,
-      paletteMode,
-      paletteOpen,
-      returnToRootPalette,
-      triggerKey,
-    ],
-  )
+  const composerPalette = useComposerPalette({
+    activePaletteIndex,
+    connectionItems,
+    disabled: composerDisabled,
+    dismissedTriggerKey,
+    dispatch: dispatchComposer,
+    draft,
+    draftSelection,
+    focusDraftAt,
+    onAddContextMention: addContextMention,
+    onViewBilling,
+    paletteMode,
+    skillItems,
+    slashItems,
+  })
 
   // 表单提交（含回车）始终走"发送"路径；"停止"只通过 ComposerTrailingControls
   // 的按钮点击触发，避免生成中按回车误中止流。
@@ -2318,105 +1214,10 @@ export function ChatArea({
       setInputError(t("chat.sendNotAccepted"))
       return
     }
-    revokeAttachmentPreviewUrls(attachments)
+    composerAttachments.revokeCurrentPreviews()
     dispatchComposer({ type: "reset-after-submit" })
     setInputError(null)
   }
-
-  const addAttachments = React.useCallback((items: AttachmentInput[]) => {
-    const next: DraftAttachment[] = []
-    for (const item of items) {
-      const attachment: DraftAttachment = {
-        id: `${Date.now()}-${item.kind ?? "file"}-${item.name}-${item.size}-${Math.random().toString(36).slice(2)}`,
-        name: item.name || item.path.split(/[\\/]/).pop() || "attachment",
-        mime: item.mime || (item.kind === "directory" ? "inode/directory" : "application/octet-stream"),
-        size: item.size,
-        path: item.path,
-        kind: item.kind ?? "file",
-      }
-      if (item.file && isImageAttachment(attachment)) {
-        attachment.previewUrl = URL.createObjectURL(item.file)
-      }
-      next.push(attachment)
-    }
-    if (next.length > 0) {
-      const existing = new Set(attachmentsRef.current.map((attachment) => attachment.path))
-      const uniqueNext = next.filter((attachment) => !existing.has(attachment.path))
-      revokeAttachmentPreviewUrls(next.filter((attachment) => existing.has(attachment.path)))
-      for (const attachment of uniqueNext) {
-        if (attachment.previewUrl) {
-          setAttachmentPreviewUrl(attachment.path, attachment.previewUrl)
-        }
-      }
-      dispatchComposer({ type: "add-attachments", attachments: uniqueNext })
-    }
-  }, [])
-
-  const addFiles = React.useCallback(
-    async (files: FileList | File[]) => {
-      setInputError(null)
-      const next: AttachmentInput[] = []
-      for (const file of Array.from(files)) {
-        const path = globalThis.lumo?.getPathForFile(file)
-        if (!path) {
-          const saver = globalThis.lumo?.saveClipboardAttachment
-          if (!saver) {
-            setInputError(t("chat.attachmentPathUnavailable"))
-            continue
-          }
-          try {
-            const saved = await saver({
-              name: file.name,
-              mime: file.type || "application/octet-stream",
-              bytes: await file.arrayBuffer(),
-            })
-            next.push({
-              name: saved.name,
-              mime: saved.mime,
-              size: saved.size,
-              path: saved.path,
-              kind: saved.kind,
-              file,
-            })
-          } catch {
-            setInputError(t("chat.attachmentSaveFailed"))
-          }
-          continue
-        }
-        next.push({
-          name: file.name || path.split(/[\\/]/).pop() || "attachment",
-          mime: file.type || "application/octet-stream",
-          size: file.size,
-          path,
-          kind: "file",
-          file,
-        })
-      }
-      addAttachments(next)
-    },
-    [addAttachments, t],
-  )
-
-  const selectAttachments = React.useCallback(
-    async (kind: "file" | "directory") => {
-      setInputError(null)
-      const picker = globalThis.lumo?.selectAttachmentPaths
-      if (!picker) {
-        if (kind === "file") {
-          fileInputRef.current?.click()
-        } else {
-          setInputError(t("chat.attachmentFolderPickerUnavailable"))
-        }
-        return
-      }
-      try {
-        addAttachments(await picker(kind))
-      } catch (error) {
-        setInputError(error instanceof Error ? error.message : String(error))
-      }
-    },
-    [addAttachments, t],
-  )
 
   const visibleError = error ?? inputError ?? modelError ?? voiceInput.error ?? voiceInput.recorderError
   const errorBanner = visibleError ? (
@@ -2431,19 +1232,8 @@ export function ChatArea({
     <PromptInput
       onSubmit={handleSubmit}
       className={cn(hasMessages && "shrink-0")}
-      onDragOver={(event) => {
-        if (!composerDisabled && event.dataTransfer.types.includes("Files")) {
-          event.preventDefault()
-        }
-      }}
-      onDrop={(event) => {
-        const files = filesFromDataTransfer(event.dataTransfer)
-        if (composerDisabled || files.length === 0) {
-          return
-        }
-        event.preventDefault()
-        void addFiles(files)
-      }}
+      onDragOver={composerAttachments.handleDragOver}
+      onDrop={composerAttachments.handleDrop}
     >
       {attachments.length > 0 || contextMentions.length > 0 ? (
         <PromptInputAttachments>
@@ -2455,14 +1245,7 @@ export function ChatArea({
             {attachments.length > 0 ? (
               <AttachmentList
                 attachments={attachments}
-                onRemove={
-                  composerDisabled
-                    ? undefined
-                    : (id) => {
-                        revokeAttachmentPreviewUrls(attachmentsRef.current.filter((attachment) => attachment.id === id))
-                        dispatchComposer({ type: "remove-attachment", id })
-                      }
-                }
+                onRemove={composerDisabled ? undefined : composerAttachments.removeAttachment}
               />
             ) : null}
           </div>
@@ -2486,36 +1269,20 @@ export function ChatArea({
             })
           }}
           onClick={updateDraftSelection}
-          onKeyDown={handleComposerKeyDown}
+          onKeyDown={composerPalette.handleKeyDown}
           onKeyUp={updateDraftSelection}
           onSelect={updateDraftSelection}
-          onPaste={(event) => {
-            const files = filesFromDataTransfer(event.clipboardData)
-            if (composerDisabled || files.length === 0) {
-              return
-            }
-            event.preventDefault()
-            void addFiles(files)
-          }}
+          onPaste={composerAttachments.handlePaste}
         />
       </PromptInputBody>
       <PromptInputToolbar>
         <PromptInputTools className="shrink-0 justify-start">
           <input
-            ref={fileInputRef}
+            ref={composerAttachments.fileInputRef}
             type="file"
             multiple
             className="hidden"
-            onChange={(event) => {
-              if (composerDisabled) {
-                event.currentTarget.value = ""
-                return
-              }
-              if (event.currentTarget.files) {
-                void addFiles(event.currentTarget.files)
-              }
-              event.currentTarget.value = ""
-            }}
+            onChange={composerAttachments.handleFileInputChange}
           />
           <DropdownMenu>
             <DropdownMenuTrigger asChild>
@@ -2532,11 +1299,11 @@ export function ChatArea({
               </Button>
             </DropdownMenuTrigger>
             <DropdownMenuContent align="start" className="min-w-40">
-              <DropdownMenuItem onSelect={() => void selectAttachments("file")}>
+              <DropdownMenuItem onSelect={() => void composerAttachments.selectAttachments("file")}>
                 <FileIcon className="size-4" />
                 {t("chat.attachFileAction")}
               </DropdownMenuItem>
-              <DropdownMenuItem onSelect={() => void selectAttachments("directory")}>
+              <DropdownMenuItem onSelect={() => void composerAttachments.selectAttachments("directory")}>
                 <Folder className="size-4" />
                 {t("chat.attachFolderAction")}
               </DropdownMenuItem>
@@ -2595,16 +1362,15 @@ export function ChatArea({
       : paletteMode === "skills"
         ? t("chat.paletteSkillsHeader")
         : undefined
-  const handlePaletteBack = activeTrigger?.kind === "slash" && paletteMode !== "root" ? returnToRootPalette : undefined
   const palette =
-    paletteOpen && activeTrigger ? (
+    composerPalette.open && composerPalette.activeTrigger ? (
       <ComposerPalette
-        activeId={activePaletteItem?.id}
+        activeId={composerPalette.activeItem?.id}
         emptyLabel={paletteEmptyLabel}
         headerLabel={paletteHeaderLabel}
-        items={paletteItems}
-        onBack={handlePaletteBack}
-        onSelect={applyPaletteItem}
+        items={composerPalette.items}
+        onBack={composerPalette.handleBack}
+        onSelect={composerPalette.onSelect}
       />
     ) : null
   const composerStack = (
